+++ src/lib/server/db/seed.ts
@@ -1,11 +1,11 @@
 import env from '$lib/env';
-import { createLogger } from '$lib/helpers/logger';
+import { createBackendLogger } from '$lib/server/logger';
 import { Table, getTableName, sql } from 'drizzle-orm';
 import { db } from './index';
 import * as schema from './schema';
 import * as seeds from './seeds';
 
-const logger = createLogger('seed');
+const logger = createBackendLogger('seed');
 
 if (!env.DB_SEEDING) {
 	throw new Error('You must set DB_SEEDING to "true" when running seeds');